Refactor TradedBondsController and QuotationsBoardContext
This commit refactors the TradedBondsController and QuotationsBoardContext classes. It includes changes to the TradedBondsController.cs file, where the code for confirming trades has been optimized and improved. It also includes changes to the QuotationsBoardContext.cs file, where the configuration for the DbContextOptions has been updated to use the appsettings.Development.json file.